refactor: enhance SymbolsTable and WatchlistsTable with improved row click handling and symbol management
