+++ src/components/TablesContainer/SymbolsTable/SymbolsTableCustomToolbar/SymbolsTableCustomToolbar.tsx
@@ -11,7 +11,7 @@ import QueryStatsIcon from '@mui/icons-material/QueryStats'
 import SearchIcon from '@mui/icons-material/Search'
 import CancelIcon from '@mui/icons-material/Cancel'
 import { ISymbolItem } from '../../../../stores/symbataStore.types.ts'
-import { useTableCustomToolbar } from './SymbolsTableCustomToolbar.hooks.ts'
+import { useTableCustomToolbar } from '../../../../hooks/useTableCustomToolbar.ts'
 
 interface ITableCustomToolbarProps {
   rows: GridRowsProp<ISymbolItem>
